requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pytest==7.4.3
httpx==0.25.2
kafka-python==2.2.15
six==1.16.0